Assistant checkpoint
Assistant generated file changes:
- templates/index.html: Add HTML interface with summarization controls
- static/style.css: Add CSS styles for the interface
- static/script.js: Add JavaScript for interface interaction
- main.py: Add new routes for summarization and model upload

---

User prompt:

Can you please give my platform an interface, and allow for there to be a adjustable scale that varies the amount of summarization anywhere from a couple of words to several paragraphs. I also want to users to be able to plug-in any model they want. Please make sure it's functional and extremely useful.